refactor: improve test organization and eliminate redundancy
- Split large combined test files into individual rule test files
- Remove duplicate test files and consolidate related tests
- Improve maintainability with dedicated test files per rule
- Maintain comprehensive test coverage while reducing redundancy
