+++ tests/test_script_magic_number_rule.py
@@ -1,24 +1,24 @@
 #!/usr/bin/env python3
-"""Unit tests for StringBooleanRule."""
+"""Unit tests for ScriptMagicNumberRule."""
 
 import pytest
-from parser.rules.structure.validation.string_boolean import StringBooleanRule
+from parser.rules.script.logic.magic_numbers import ScriptMagicNumberRule
 from parser.models import ProjectContext
 
 
-class TestStringBooleanRule:
-    """Test cases for StringBooleanRule class."""
+class TestScriptMagicNumberRule:
+    """Test cases for ScriptMagicNumberRule class."""
     
     def setup_method(self):
         """Set up test fixtures."""
-        self.rule = StringBooleanRule()
+        self.rule = ScriptMagicNumberRule()
         self.context = ProjectContext()
     
     def test_rule_metadata(self):
         """Test rule metadata is correctly set."""
         assert self.rule.ID == "RULE000"  # Base class default
         assert self.rule.SEVERITY == "ADVICE"
-        assert "string" in self.rule.DESCRIPTION.lower()
+        assert "magic" in self.rule.DESCRIPTION.lower()
 
 
 if __name__ == '__main__':
